+++ inveroryManagerApp/Views/Customers/BuyItems.cshtml
