+++ inveroryManagerApp/Views/Customers/ViewTransactions.cshtml
