+++ inveroryManagerApp/Views/Home/Customers.cshtml
@@ -6,19 +6,19 @@
 </div>
 
 <div class="position-absolute top-50 start-50 translate-middle ">
-    <a asp-action="ListCustomers" asp-controller="home" class="btn btn-primary btn-lg"
+    <a asp-action="ListCustomers" asp-controller="Customers" class="btn btn-primary btn-lg"
        style="height: max-content width: max-content; font-size: 2vw;">
         List Customers
     </a>
-    <a asp-action="AddCustomer" asp-controller="home" class="btn btn-primary btn-lg"
+    <a asp-action="AddCustomer" asp-controller="Customers" class="btn btn-primary btn-lg"
        style="height: max-content; width: max-content; font-size: 2vw;">
         Add Customer
     </a>
-    <a asp-action="BuyItems" asp-controller="home" class="btn btn-primary btn-lg"
+    <a asp-action="BuyItems" asp-controller="Customers" class="btn btn-primary btn-lg"
        style="height: max-content; width: max-content; font-size: 2vw;">
         Buy Items
     </a>
-    <a asp-action="ListTransactions" asp-controller="home" class="btn btn-primary btn-lg"
+    <a asp-action="ListTransactions" asp-controller="Customers" class="btn btn-primary btn-lg"
        style="height: max-content; width: max-content; font-size: 2vw;">
         List Transactions
     </a>
