+++ inveroryManagerApp/Views/Home/Staffs.cshtml
@@ -6,11 +6,11 @@
 </div>
 
 <div class="position-absolute top-50 start-50 translate-middle ">
-    <a asp-action="ListStaffs" asp-controller="home" class="btn btn-primary btn-lg"
+    <a asp-action="ListStaffs" asp-controller="Staff" class="btn btn-primary btn-lg"
        style="height: max-content width: max-content; font-size: 2vw;">
         List Staffs
     </a>
-    <a asp-action="AddStaff" asp-controller="home" class="btn btn-primary btn-lg"
+    <a asp-action="AddStaff" asp-controller="Staff" class="btn btn-primary btn-lg"
        style="height: max-content; width: max-content; font-size: 2vw;">
         Add Staff
     </a>
